flask_api/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
numpy>=2.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai>=0.3.0
xarray>=2023.0.0
h5netcdf>=1.2.0
